Add trait for language detector
