+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/serialization/KryoRegistrator.scala
@@ -24,7 +24,8 @@ class KryoRegistrator extends BaseKryoRegistrator {
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.image.ImageAnalyzer])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.image.ImageAnalysisAuth])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.language.LanguageDetector])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.language.LanguageDetectorAuth])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.language.CognitiveServicesLanguageDetector])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.language.CognitiveServicesLanguageDetectorAuth])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.sentiment.SentimentDetector])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.sentiment.CognitiveServicesSentimentDetector])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.sentiment.WordListSentimentDetector])
