+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/transformcontext/Delta.scala
@@ -2,7 +2,7 @@ package com.microsoft.partnercatalyst.fortis.spark.transformcontext
 
 import com.microsoft.partnercatalyst.fortis.spark.dto.{BlacklistedTerm, SiteSettings}
 import com.microsoft.partnercatalyst.fortis.spark.transforms.image.{ImageAnalysisAuth, ImageAnalyzer}
-import com.microsoft.partnercatalyst.fortis.spark.transforms.language.{LanguageDetector, LanguageDetectorAuth}
+import com.microsoft.partnercatalyst.fortis.spark.transforms.language.{CognitiveServicesLanguageDetector, CognitiveServicesLanguageDetectorAuth, LanguageDetector}
 import com.microsoft.partnercatalyst.fortis.spark.transforms.locations.LocationsExtractorFactory
 import com.microsoft.partnercatalyst.fortis.spark.transforms.locations.client.FeatureServiceClient
 import com.microsoft.partnercatalyst.fortis.spark.transforms.sentiment.SentimentDetectorAuth
@@ -63,7 +63,7 @@ private[transformcontext] object Delta {
       ),
       languageDetector = updatedField(
         siteSettings.get.translationSvcToken != transformContext.siteSettings.translationSvcToken,
-        new LanguageDetector(LanguageDetectorAuth(siteSettings.get.translationSvcToken))
+        new CognitiveServicesLanguageDetector(CognitiveServicesLanguageDetectorAuth(siteSettings.get.translationSvcToken))
       ),
       sentimentDetectorAuth = updatedField(
         siteSettings.get.translationSvcToken != transformContext.siteSettings.translationSvcToken,
