+++ src/test/scala/com/microsoft/partnercatalyst/fortis/spark/transforms/language/CognitiveServicesLanguageDetectorSpec.scala
@@ -2,15 +2,15 @@ package com.microsoft.partnercatalyst.fortis.spark.transforms.language
 
 import org.scalatest.FlatSpec
 
-class TestLanguageDetector(cognitiveServicesResponse: String) extends LanguageDetector(LanguageDetectorAuth("key")) {
+class TestCognitiveServicesLanguageDetector(cognitiveServicesResponse: String) extends CognitiveServicesLanguageDetector(CognitiveServicesLanguageDetectorAuth("key")) {
   protected override def callCognitiveServices(request: String): String = cognitiveServicesResponse
   override def buildRequestBody(text: String, textId: String): String = super.buildRequestBody(text, textId)
 }
 
-class LanguageDetectorSpec extends FlatSpec {
+class CognitiveServicesLanguageDetectorSpec extends FlatSpec {
   "The language detector" should "formulate correct request and parse response to domain types" in {
     val responseConfidence = 1.0
-    val detector = new TestLanguageDetector(s"""{"documents":[{"id":"0","detectedLanguages":[{"name":"English","iso6391Name":"en","score":$responseConfidence}]}],"errors":[]}""")
+    val detector = new TestCognitiveServicesLanguageDetector(s"""{"documents":[{"id":"0","detectedLanguages":[{"name":"English","iso6391Name":"en","score":$responseConfidence}]}],"errors":[]}""")
     val language = detector.detectLanguage("some text")
 
     assert(language.contains("en"))
@@ -18,7 +18,7 @@ class LanguageDetectorSpec extends FlatSpec {
 
   it should "ignore low confidence detections" in {
     val responseConfidence = 0.01
-    val detector = new TestLanguageDetector(s"""{"documents":[{"id":"0","detectedLanguages":[{"name":"English","iso6391Name":"en","score":$responseConfidence}]}],"errors":[]}""")
+    val detector = new TestCognitiveServicesLanguageDetector(s"""{"documents":[{"id":"0","detectedLanguages":[{"name":"English","iso6391Name":"en","score":$responseConfidence}]}],"errors":[]}""")
     val language = detector.detectLanguage("some text")
 
     assert(language.isEmpty)
@@ -27,7 +27,7 @@ class LanguageDetectorSpec extends FlatSpec {
   it should "build correct request body" in {
     val id = "0"
     val text = "some text"
-    val requestBody = new TestLanguageDetector("").buildRequestBody(text, id)
+    val requestBody = new TestCognitiveServicesLanguageDetector("").buildRequestBody(text, id)
 
     assert(requestBody == s"""{"documents":[{"id":"$id","text":"$text"}]}""")
   }
